+++ tests/testthat/test-municipalities.R
@@ -1,30 +1,37 @@
-# Test municipality data: Jilove_DC_cz
+# Municipality datasets tests
 
-test_that("Jilove_DC_cz data is a valid election dataset", {
-
-  expect_s3_class(Jilove_DC_cz, "data.frame")
-
+test_that("municipality datasets are valid election datasets", {
+  
+  municipalities_list <- list(Doubice_DC_cz = Doubice_DC_cz,
+                              Jilove_DC_cz  = Jilove_DC_cz);
+  
   expected_cols <- c("elections", "candidate", "list_name", "list_pos",
                      "pref_votes", "elected", "nom_party", "pol_affil",
                      "mayor", "dep_mayor", "board", "gov_support",
                      "elig_voters", "ballots_cast")
-
-  expect_named(Jilove_DC_cz, expected_cols, ignore.order = TRUE)
-  expect_true(nrow(Jilove_DC_cz) > 0)
-
-  expect_type(Jilove_DC_cz$elections,    "integer")
-  expect_type(Jilove_DC_cz$candidate,    "character")
-  expect_type(Jilove_DC_cz$list_name,    "character")
-  expect_type(Jilove_DC_cz$list_pos,     "integer")
-  expect_type(Jilove_DC_cz$pref_votes,   "integer")
-  expect_type(Jilove_DC_cz$elected,      "integer")
-  expect_type(Jilove_DC_cz$nom_party,    "character")
-  expect_type(Jilove_DC_cz$pol_affil,    "character")
-  expect_type(Jilove_DC_cz$mayor,        "integer")
-  expect_type(Jilove_DC_cz$dep_mayor,    "integer")
-  expect_type(Jilove_DC_cz$board,        "integer")
-  expect_type(Jilove_DC_cz$gov_support,  "integer")
-  expect_type(Jilove_DC_cz$elig_voters,  "integer")
-  expect_type(Jilove_DC_cz$ballots_cast, "integer")
-
-}) # konce testu pro Jilove_DC_cz
+  
+  for (municipality in municipalities_list) {
+    
+    expect_s3_class(municipality, "data.frame");
+    
+    expect_named(municipality, expected_cols, ignore.order = TRUE)
+    expect_true(nrow(municipality) > 0)
+    
+    expect_type(municipality$elections,    "integer")
+    expect_type(municipality$candidate,    "character")
+    expect_type(municipality$list_name,    "character")
+    expect_type(municipality$list_pos,     "integer")
+    expect_type(municipality$pref_votes,   "integer")
+    expect_type(municipality$elected,      "integer")
+    expect_type(municipality$nom_party,    "character")
+    expect_type(municipality$pol_affil,    "character")
+    expect_type(municipality$mayor,        "integer")
+    expect_type(municipality$dep_mayor,    "integer")
+    expect_type(municipality$board,        "integer")
+    expect_type(municipality$gov_support,  "integer")
+    expect_type(municipality$elig_voters,  "integer")
+    expect_type(municipality$ballots_cast, "integer")
+    
+  } # konec FOR loopu
+  
+}) # konce testu pro municipalities